autotest_package/requirements.txt
selenium>=4.0.0
webdriver-manager
requests
python-dotenv
jsonschema
pyyaml
pdfplumber
pypdf2
anthropic
openai
pytest
jinja2
flake8
langchain
langchain-anthropic
langchain-google-genai
google-generativeai
langchain-mistralai
langchain-ollama
langchain-groq
python-dotenv
